+++ src/cwprep/__init__.py
@@ -28,7 +28,7 @@ from .config import (
     DEFAULT_CONFIG
 )
 
-__version__ = "0.5.0"
+__version__ = "0.5.1"
 __author__ = "cooper wenhua"
 __all__ = [
     "TFLBuilder",